Removed console logs, and regactored executeCode route to a controller
